requirements.txt
flask
flask_cors
numpy
torch
torchvision
timm
albumentations
opencv-python-headless==4.12.0.88
nibabel
catboost
xgboost
scikit-learn==1.6.1
Pillow>=9.0.0
tqdm>=4.64.0
